fixes in general and ZMoves
