+++ ZTeraMax/config.yml
@@ -5,7 +5,7 @@ authors:
   - Aelysya
   - Lexio
   - Ota
-version: 0.8.34
+version: 0.8.55
 deps: []
 retry_psdk_compatibility_after_plugin_load: false
 added_files:
