+++ ZTeraMax/scripts/0 General/004 PokemonBattler.rb
@@ -1,5 +1,4 @@
 module PFM
-  # Class defining a Pokemon during a battle, it aim to copy its properties but also to have the methods related to the battle.
   class PokemonBattler < Pokemon
     module PokemonBattlerZTeraMaxPlugin
       # @return [Array<Battle::Move>] the original moveset of the Pokemon
@@ -18,7 +17,6 @@ module PFM
         @original_moveset = []
       end
     end
-
     prepend PokemonBattlerZTeraMaxPlugin
   end
 end
