+++ ZTeraMax/scripts/1 ZMoves/007 CrystalItemsGiveTake.rb
@@ -1,22 +1,17 @@
 module Battle
   class Logic
     class ItemChangeHandler < ChangeHandlerBase
-      module ZMovesPlugin
-        # List of item that cannot be manipulated during battle (Knock off, Trick, etc...)
-        PROTECTED_ITEMS.concat(%i[normalium_z fightinium_z flyinium_z poisonium_z groundium_z rockium_z buginium_z
-                                  ghostium_z steelium_z firium_z waterium_z grassium_z electrium_z psychium_z
-                                  icium_z dragonium_z darkinium_z fairium_z aloraichium_z decidium_z eevium_z
-                                  incinium_z kommonium_z lunalium_z lycanium_z marshadium_z mewnium_z mimikium_z
-                                  pikanium_z pikashunium_z primarium_z snorlium_z solganium_z tapunium_z ultranecrozium_z])
-      end
-
-      prepend ZMovesPlugin
+      # List of item that cannot be manipulated during battle (Knock off, Trick, etc...)
+      PROTECTED_ITEMS.concat(%i[normalium_z fightinium_z flyinium_z poisonium_z groundium_z rockium_z buginium_z
+                                ghostium_z steelium_z firium_z waterium_z grassium_z electrium_z psychium_z
+                                icium_z dragonium_z darkinium_z fairium_z aloraichium_z decidium_z eevium_z
+                                incinium_z kommonium_z lunalium_z lycanium_z marshadium_z mewnium_z mimikium_z
+                                pikanium_z pikashunium_z primarium_z snorlium_z solganium_z tapunium_z ultranecrozium_z])
     end
   end
 end
 
 module Util
-  # Module adding the give / take item functionality to a scene
   module GiveTakeItem
     # List of Z crystals IDs associations, key is the big crystal, value is the held crystal
     Z_CRYSTALS_IDS = {
