+++ ZTeraMax/scripts/1 ZMoves/1 Moves Definitions/003 StatusZMove.rb
