+++ ZTeraMax/scripts/1 ZMoves/2 Moves Modification/300 NaturePower.rb
@@ -18,7 +18,7 @@ module Battle
           log_data("nature power # becomes #{skill.db_symbol}")
 
           move = Battle::Move[skill.be_method].new(skill.db_symbol, 1, 1, @scene)
-          move = logic.z_move.corresponding_z_move(move) unless move.status?
+          move = logic.z_move.get_corresponding_z_move(move) unless move.status?
           def move.usage_message(user)
             @scene.visual.hide_team_info
             scene.display_message_and_wait(parse_text(18, 127, '[VAR MOVE(0000)]' => name))
